docker/database/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
weaviate-client==3.25.3
sentence-transformers==2.7.0
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
numpy==1.24.4
torch==2.1.1
transformers==4.36.0
aiofiles==23.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
huggingface_hub==0.19.4 